fixed problem with sorter (product-service)

+++ product-service/src/main/java/com/example/productservice/models/DTO/ProductSorter.java
@@ -4,7 +4,6 @@ import lombok.Data;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Size;
-import java.math.BigDecimal;
 
 @Data
 public class ProductSorter {
@@ -12,8 +11,8 @@ public class ProductSorter {
     private String title;
 
     @Max(Integer.MAX_VALUE)
-    private BigDecimal priceFrom;
+    private Integer priceFrom;
 
     @Max(Integer.MAX_VALUE)
-    private int priceTo;
+    private Integer priceTo;
 }
